Conflicts:
	ATmega Firmata/1sheelds_functions.c
	ATmega Firmata/mapping162.h

+++ ATmega Firmata/1sheelds_functions.h
@@ -10,11 +10,17 @@
   Date:          2014.5
 
 */
+#include "sys.h"
+#include "gpio.h"
+#include "pwm.h"
+#include "uart.h"
+#include "mapping162.h"
+#include <util/delay.h>
 
 #ifndef SHEELDS_FUNCTIONS_H_
 #define SHEELDS_FUNCTIONS_H_
 
-#include "sys.h"
+
 
 uint8  digitalRead(uint8);
 void   digitalWrite(uint8, uint8);
@@ -24,7 +30,7 @@ unsigned long pulseIn(uint8_t pin, uint8_t state);
 unsigned long readPWM (int pin);
 void   writePort(byte, byte, byte);
 uint8  readPort(byte, byte);
-uint16_t serial0_Avilable();
+int serial0_Avilable();
 int serial1_Avilable();
 void turnOffPWM(uint8 timer);
 void unusedPinsAsOutput();
